feat(web): relay / network parity on web
