+++ engine/protocol/src/lib.rs
@@ -1,9 +1,14 @@
 pub mod receive;
+pub mod relay;
 pub mod send;
 pub mod time_compat;
 pub mod types;
 
 pub use receive::{download_to_store, fetch_metadata, DownloadToStoreResult};
+pub use relay::{
+    build_relay_mode, get_relay_status, relay_fallback_policy, resolve_relay_mode_with_fallback,
+    verify_relays, RelayConfigArg, RelayFallbackPolicy, RelayStatusResponse, VerifyRelaysResponse,
+};
 pub use send::{run_share_session, ShareSessionOutcome, METADATA_ALPN};
 pub use types::*;
